Add Content in Course Setting
Now it works for all content types and you can switch between them. If
adding content from outside a course, the query string will end with a
zero indicating that the content not be added to a course. The redirect
will be to the view content page rather than to the course page.